+++ NEWS.md
@@ -1,6 +1,6 @@
 # scales 1.0.0.9000
 
-* `parse_format()` now returns an expression object which can be used to display formatted labels in ggplot2 (@agila5, #203)
+* `parse_format()` now returns a function which returns expression object that can be used to display formatted labels in ggplot2 (@agila5, #203)
 
 * `boxcox_trans()` no longer throws an error when given NA values (@sflippl, #181)